+++ Applications/FileManager/DirectoryView.h
@@ -101,7 +101,7 @@ private:
     ViewMode m_view_mode { Invalid };
 
     NonnullRefPtr<GUI::FileSystemModel> m_model;
-    int m_path_history_position { 0 };
+    size_t m_path_history_position { 0 };
     Vector<String> m_path_history;
     void add_path_to_history(const StringView& path);
